Refactor endpoint rule reference helpers
